requirements.txt
# Core Deep Learning and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Data Processing and Augmentation
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
albumentations>=1.3.0
imgaug>=0.4.0

# Roboflow Integration
roboflow>=1.1.0

# Visualization and Plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Model Evaluation and Metrics
scikit-learn>=1.3.0
scipy>=1.11.0

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Progress Bars and Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter Notebook Support
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# API and Web Framework (for future extensions)
flask>=2.3.0
flask-cors>=4.0.0

# File Management
pathlib>=1.0.1
glob2>=0.7

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Optional GPU Acceleration
# cudatoolkit=11.8  # Install via conda if using CUDA
# cudnn=8.6.0       # Install via conda if using CUDA

# System Utilities
psutil>=5.9.0
requests>=2.31.0
